Configure the GCode file name format for better versioning
By default the format contains variable stuff that would result in carrying multiple versions of the file side by side, instead of carrying a history of a single file. By changing the format to use just the file name, this is avoided.

This is configured in Print Settings (Expert) > Output file > Output filename format.