Change to a fit testing box for the new layout idea
The monitor mount has two parts which connect using a track for one to go into the other and are secured together using screws. I did not have these screws there and tightened before. I added them and now the distance between the mount plate and the wall is much shorter so the previous revision doesn't fit at all. I've decided to change course and instead print a stand that will only attach to the mount part that is secured to the monitor not primarily the one attached to the wall.