Update dimensions and form into a hook platform over the VESA mount
I need to learn to use the calipers properly, because the previous fit test box was too short on the X (but the other dimensions were correct).

This shape should be able to sit snugly on the VESA mount edge and secured by the two screws so it doesn't move in any direction.