Revert the GCode file name config
I tend to start a new PrusaSlicer project and just delete the 3MF directory in this repository, save the new project and rebuild it.

This setting is not preserved when I create a new project and I have found it is not too much of an issue since I can just select the old GCode file when exporting GCode and it will take its name and replace it, so the default suggested name is no problem. The modal window even remembers the last directory, so it is quite literally only a single click to solve this problem.

I could probably set this in some sort of a profile file and have it be the default for new prints, but I don't want that for prints that are not versioned in Git, one-offs, because I keep a history of those GCode files on my SD card and I actually like versioning them by basically date and print time in the directory there, so this would not be a good global default.