Split the readme into Status and Process sections
The process section is important so that I can remember this workflow if I ever get back to this at a later date.

The status section captures the goal as well as the current state of the part as it is undergoing development. Once done, I will reflect that in this section.