Redesign the rectangular slot that goes between the screen and the mount
I have messed up before, the screw slots need to be at the right distance from the tip of the plate arc, not right beneath its level. I've reduced the design to just this slot for now and I will tweak it until it is good and then continue designing for the other part of the plate.