Increase screw diameter for clearance and add opposite side rectangular leg
The screw slot is too tight because I made it the exact screw diameter and the nozzle diameter is not a perfect division of that number. I added an extra give hoping to accomodate the screws better.

I also added a draft slot for the leg on the opposite side of the VESA mount plate.