Increase screw diameter for even more clearance and shrink back plate
The screw slots still do not fit snugly around the screw so I've bumped the clearance up even more. On top of this, I have shrunk the back plate to make sure it does not collide with the flaps coming up from the plate, which would probably not happen with its current depth but would have happened if I prolonged it later.