Print plate-screen gap depth test fit piece in vase mode
I know the dimensions of the gap in two axes which give me the rectangular main leg footprint, however, I don't know how deep this gap can go, so I'm trying to see if 100 mm is too much or will still fit.

I have measured the distance between the flaps on the mount plate, it is 75 mm as compared to the distance between the screws, which is 96 mm. This means the main leg needs to be at most 75 mm thick, not 96 mm, after a certain depth (the distance between the flaps and the screws). I have not measured this exact depth yet, will do so after the test fit.

I'm printing with no top or bottom solid layers or infill and in vase mode so I get a super speedy print.