Increase flap distance to make the model fit
Previous test fit did not go accoding to plan, it seems the value for the flap distance I measured on the other side of the plate previously when making the other, currently hidden flat rectangular leg, is a better fit.

I might adjust this even more if this still won't fit.