Add the other rectangular leg to test fit both at the same time
The width of the rectangular leg is 74 mm as verified by the last test fit. I am winding a 1.6 mm mount plate thickness as 2 mm seemed to have too much give. I've also made the other rectangular leg go a bit deeper into the mount edge, but left room so that it would for sure fit. This whole assembly should fit now. If it does, it will be time to add the spacer part and the top flat layer to sit the camera on to it.