Reorient for slower printing but without irremovable supports
I made a mistake trying to print the part laying flat before, because the support in the thin slot was impossible to remove and I permanently bent the part with still some support left trying to do it. I will print in a standing orientation now and to prevent layer shifting and other bad stuff, I have increased perimeter enough so that only the base will have infill and the rest of the model will be solid inside. This should give it enough strength to print cleanly and last. I have also added a lot of brim hoping it would help keep the object steady while printing.