Simplify the shape of the main leg for faster printing
This triangular shape will make it easier to fit between the screws and the flaps and the bridge surfaces over the screws and the flaps should still keep the piece level and sturdy. Maybe this triangular part could be forgone completely and sturdiness could be derived solely from the fit over the edge of the mount plate, but I don't want to rely on that alone yet as I am still experimenting with the plate thickness which will make the piece fit easily but not have too much play.

I have not updated the 3MF and the GCode as I am not printing this, I will adjust the plate thickness in the next commit and do a test print then.