Increase gap for plate side mounting
I have removed the dimension to make this part slottable into another, I am thinking I will print this as a single part after all since the gap depth does not need to extend all so much.

I have updated PrusaSlicer which changed the settings diff a little bit, but that's unrelated to the geometry change.