Shift and enlarge top surface and add support brace
I think I have the height sort of okay, but the previous extension of the surface to both sides made it crash to the wall and also I misplaced the dummy battery plug hole. The hole would still have been okay, just weirdly placed, but the crashing of the surface to the wall caused the whole assembly to not be able to fit onto the edge of the mount plate.

It was also too narow so the camera would cause it to tip forward and since I had to break off the support flap on the back, it would not resist because it was not holding onto both sides of the mount plate.

I was not sure whether the surface would be able to hold the weight of the camera so I added a support brace which should be able to withstand it.

I am not printing top layers (in print orientation - in actual orientation, bottom layers) as they will not be seen when mounted so I am saving some print time this way.