Fix the dummy battery plug hole dimensions
I goofed out and placed the hole over the area where the plug cover rotates over to when the camera plug is exposed, but the area where the plug hole and the plug on the cable connect together was not accounted for. Also the cable and the plug on it need room to bend out of the way. I made the slit wide enough to accomodate for all this now - I hope.